+++ src/OpenRasta/OperationModel/MethodBased/AsyncOperationDescriptor.cs
@@ -1,0 +1,12 @@
+using System;
+using OpenRasta.TypeSystem;
+
+namespace OpenRasta.OperationModel.MethodBased
+{
+  public class AsyncOperationDescriptor : OperationDescriptor
+  {
+    public AsyncOperationDescriptor(IMethod method, Func<IOperationAsync> factory) : base(method, factory)
+    {
+    }
+  }
+}
